Add ability to access flags from Transaction.java (#690)
* Introduce `transactionFlags()` in a manner that doesn’t break immutables.
* Add unit tests to all transactions to verify functionality.
* Note this addition is required for Batch functionality to verify inner transaction flags in a generic manner.

+++ xrpl4j-core/src/test/java/org/xrpl/xrpl4j/model/client/fees/FeeUtilsTest.java
@@ -9,9 +9,9 @@ package org.xrpl.xrpl4j.model.client.fees;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -445,7 +445,7 @@ public class FeeUtilsTest {
 
     assertThrows(
       NullPointerException.class,
-      () -> FeeUtils.min(BigInteger.ZERO, null)
+      () -> FeeUtils.min(BigInteger.ZERO, (BigInteger) null)
     );
 
     assertThat(FeeUtils.min(BigInteger.valueOf(-1), BigInteger.ZERO)).isEqualTo(BigInteger.valueOf(-1));
@@ -467,7 +467,7 @@ public class FeeUtilsTest {
 
     assertThrows(
       NullPointerException.class,
-      () -> FeeUtils.max(BigInteger.ZERO, null)
+      () -> FeeUtils.max(BigInteger.ZERO, (BigInteger) null)
     );
 
     assertThat(FeeUtils.max(BigInteger.valueOf(-1), BigInteger.ZERO)).isEqualTo(BigInteger.ZERO);
